trainer/requirements.txt
tensorflow==2.3.0
numpy==1.18.5
Pillow==7.2.0
requests==2.24.0
google-cloud-storage==1.16.1
tweepy==3.9.0
scikit-learn==0.23.2
matplotlib==3.3.1
astral==2.2.0
google-api-python-client==1.12.5
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
